python-pdf-imagem/requirements.txt
Pillow==3.1.1
reportlab==3.3.0
